abort running mocha tests when webpack starts to compile (#94)
When webpack detects some file changes the currently running test suite will be aborted and after that webpack continues to compile. This ensures that everything gets executed in the right order and that makes building a logging system a lot easier. The most important information is always the last one and not mixed like before.