use "unique" output path for bundle (#165)
Normally the output bundle will be written to and read from memory fs but if the file also exists on the real fs the real fs takes precedence. We need to work around this with an "unique" output path to ensure that the file never exists on file system.

This is also useful when multiple instances of mocha-webpack are running parallel.